Shifting routesPostMethod to reduce its access
