+++ sdk-springboot/src/main/java/io/dapr/springboot/DaprBeanPostProcessor.java
@@ -17,7 +17,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.dapr.Topic;
-import io.dapr.springboot.util.SpringProcessorHelper;
 import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
@@ -29,6 +28,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -92,7 +92,7 @@ public class DaprBeanPostProcessor implements BeanPostProcessor {
           TypeReference<HashMap<String, String>> typeRef
                   = new TypeReference<HashMap<String, String>>() {};
           Map<String, String> metadata = MAPPER.readValue(topic.metadata(), typeRef);
-          List<String> routes = SpringProcessorHelper.getAllCompleteRoutesForPost(clazz, method, topicName);
+          List<String> routes = getAllCompleteRoutesForPost(clazz, method, topicName);
           for (String route : routes) {
             DaprRuntime.getInstance().addSubscribedTopic(pubSubName, topicName, route,
                 metadata);
@@ -105,5 +105,71 @@ public class DaprBeanPostProcessor implements BeanPostProcessor {
     }
   }
 
+  /**
+   * Method to provide all possible complete routes list fos this post method present in this controller class,
+   * for mentioned topic.
+   *
+   * @param clazz     Controller class
+   * @param method    Declared method for posting data
+   * @param topicName Associated topic name
+   * @return All possible routes for post mapping for this class and post method
+   */
+  private static List<String> getAllCompleteRoutesForPost(Class clazz, Method method, String topicName) {
+    List<String> routesList = new ArrayList<>();
+    RequestMapping clazzRequestMapping =
+        (RequestMapping) clazz.getAnnotation(RequestMapping.class);
+    String[] clazzLevelRoute = null;
+    if (clazzRequestMapping != null) {
+      clazzLevelRoute = clazzRequestMapping.value();
+    }
+    String[] postValueArray = getRoutesForPost(method, topicName);
+    if (postValueArray != null && postValueArray.length >= 1) {
+      for (String postValue : postValueArray) {
+        if (clazzLevelRoute != null && clazzLevelRoute.length >= 1) {
+          for (String clazzLevelValue : clazzLevelRoute) {
+            String route = clazzLevelValue + confirmLeadingSlash(postValue);
+            routesList.add(route);
+          }
+        } else {
+          routesList.add(postValue);
+        }
+      }
+    }
+    return routesList;
+  }
+
+  private static String[] getRoutesForPost(Method method, String topicName) {
+    String[] postValueArray = new String[] {topicName};
+    PostMapping postMapping = method.getAnnotation(PostMapping.class);
+    if (postMapping != null) {
+      if (postMapping.path() != null && postMapping.path().length >= 1) {
+        postValueArray = postMapping.path();
+      } else if (postMapping.value() != null && postMapping.value().length >= 1) {
+        postValueArray = postMapping.value();
+      }
+    } else {
+      RequestMapping reqMapping = method.getAnnotation(RequestMapping.class);
+      for (RequestMethod reqMethod : reqMapping.method()) {
+        if (reqMethod == RequestMethod.POST) {
+          if (reqMapping.path() != null && reqMapping.path().length >= 1) {
+            postValueArray = reqMapping.path();
+          } else if (reqMapping.value() != null && reqMapping.value().length >= 1) {
+            postValueArray = reqMapping.value();
+          }
+          break;
+        }
+      }
+    }
+    return postValueArray;
+  }
+
+  private static String confirmLeadingSlash(String path) {
+    if (path != null && path.length() >= 1) {
+      if (!path.substring(0, 1).equals("/")) {
+        return "/" + path;
+      }
+    }
+    return path;
+  }
 
 }
